+++ src/main/kotlin/blackjack/domain/BetAmount.kt
@@ -1,7 +1,5 @@
 package blackjack.domain
 
-import blackjack.domain.PlayerResultStatus.*
-
 class BetAmount(private var amount: Int) {
     init {
         require(amount > 0) { "배팅 금액은 0원 이상 입력해주세요." }
