+++ src/main/kotlin/blackjack/domain/participant/PlayerResultStatus.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.participant
 
 enum class PlayerResultStatus {
     BLACKJACK_WIN,
